refact[4.2]: move logic from index.js to blogs controller and model
