+++ controllers/blogs.js
@@ -1,0 +1,30 @@
+const blogsRouter = require('express').Router()
+const Blog = require('../models/Blog')
+
+blogsRouter.get('/api/blogs', (request, response) => {
+  Blog
+    .find({})
+    .then(blogs => {
+      response.json(blogs)
+    })
+})
+
+blogsRouter.get('/api/blogs/:id', (request, response) => {
+  Blog
+    .findById(request.params.id)
+    .then(blog => {
+      response.json(blog)
+    })
+})
+
+blogsRouter.post('/api/blogs', (request, response) => {
+  const blog = new Blog(request.body)
+
+  blog
+    .save()
+    .then(result => {
+      response.status(201).json(result)
+    })
+})
+
+module.exports = blogsRouter
